requirements.txt
hidapi
pystray
pillow
pycaw
comtypes
win10toast
pyinstaller
